% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_altmetric}
\alias{badge_altmetric}
\title{badge_altmetric}
\usage{
badge_altmetric(id, color, alt = NULL)
}
\arguments{
\item{id}{altmetric id}

\item{color}{color of badge}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
altmetric badge
}
\examples{
\dontrun{
badge_altmetric("2788597", "blue")
}
}
\author{
Guangchuang Yu
}
