% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_bioc_download}
\alias{badge_bioc_download}
\title{badge_bioc_download}
\usage{
badge_bioc_download(pkg = NULL, by, color, type = "distinct", alt = NULL)
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}

\item{by}{one of total or month}

\item{color}{badge color}

\item{type}{one of distinct and total}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
badge of bioconductor download number
}
\author{
Guangchuang Yu
}
