% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_bioc_release}
\alias{badge_bioc_release}
\title{badge_bioc_release}
\usage{
badge_bioc_release(pkg = NULL, color, alt = NULL)
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}

\item{color}{badge color}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
badge of bioconductor release version
}
\author{
Guangchuang Yu
}
