% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_codefactor}
\alias{badge_codefactor}
\title{CodeFactor badge}
\usage{
badge_codefactor(ref = NULL, alt = "CodeFactor")
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL} (the default),
the reference is determined by the URL field in the DESCRIPTION file.}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
CodeFactor badge
}
\author{
Alexander Rossell Hayes
}
