% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_custom}
\alias{badge_custom}
\title{badge_custom}
\usage{
badge_custom(x, y, color, url = NULL, alt = NULL)
}
\arguments{
\item{x}{field 1}

\item{y}{field 2}

\item{color}{color}

\item{url}{optional url}

\item{alt}{Alternative text}
}
\value{
customize badge
}
\description{
custom badge
}
\author{
Guangchuang Yu
}
