% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_download_bioc}
\alias{badge_download_bioc}
\title{badge_download_bioc}
\usage{
badge_download_bioc(pkg)
}
\arguments{
\item{pkg}{package}
}
\value{
bioc download badge
}
\description{
official Bioconductor download badge
}
\author{
guangchuang yu
}
