% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_github_check}
\alias{badge_github_check}
\title{GitHub Actions R-CMD-CHECK badge}
\usage{
badge_github_check(ref = NULL)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL} (the default),
the reference is determined by the URL field in the DESCRIPTION file.}
}
\value{
badge in Markdown syntax
}
\description{
GitHub Actions R-CMD-CHECK badge
}
\author{
Alexander Rossell Hayes
}
