% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_lifecycle}
\alias{badge_lifecycle}
\title{badge_lifcycle}
\usage{
badge_lifecycle(stage = "experimental", color = NULL, alt = NULL)
}
\arguments{
\item{stage}{lifecycle stage See
\href{https://lifecycle.r-lib.org/articles/stages.html}{https://lifecycle.r-lib.org/articles/stages.html}}

\item{color}{color of the badge. If missing, the color is determined by the stage.}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
lifecycle badge
}
\author{
Gregor de Cillia, Waldir Leoncio
}
