% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_repostatus}
\alias{badge_repostatus}
\title{badge_repostatus}
\usage{
badge_repostatus(status, alt = NULL)
}
\arguments{
\item{status}{project status (Concept, WIP, Suspended, Abandoned, Active,
Inactive, Unsupported, Moved). See
\href{https://www.repostatus.org/lifecycle}{https://www.repostatus.org/lifecycle}
for more details.}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
repostatus.org badge
}
\author{
Waldir Leoncio
}
