% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_sci_citation}
\alias{badge_sci_citation}
\title{badge_sci_citation}
\usage{
badge_sci_citation(url, color, alt = NULL)
}
\arguments{
\item{url}{Web of Science url}

\item{color}{color of badge}

\item{alt}{Alternative text}
}
\value{
badge in markdown syntax
}
\description{
SCI citation badge
}
\author{
Guangchuang Yu
}
