% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_bioc_download_rank}
\alias{badge_bioc_download_rank}
\alias{badge_download_bioc}
\title{badge_bioc_download_rank}
\usage{
badge_bioc_download_rank(pkg = NULL, alt = "download")

badge_download_bioc(pkg = NULL, alt = "download")
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}

\item{alt}{Alternative text}
}
\value{
bioc download ranking badge
}
\description{
official Bioconductor download badge (download ranking)
}
\author{
Guangchuang Yu
}
