% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpModel.R
\name{calcAlleleCnvProb}
\alias{calcAlleleCnvProb}
\title{Run BADGER allele-only model to assess posterior probability of CNVs given allele information only}
\usage{
calcAlleleCnvProb(r, cov.sc, l, cov.bulk, region, gtf, mono = 0.7,
  pe = 0.01, filter = TRUE, likelihood = FALSE, n.iter = 1000,
  quiet = TRUE, delim = ":")
}
\arguments{
\item{r}{Matrix of alt allele count in single cells}

\item{cov.sc}{Matrix of coverage in single cells}

\item{l}{Vector of alt allele count in bulk}

\item{cov.bulk}{Vector of coverage in bulk}

\item{region}{Region of interest such as expected CNV boundaries}

\item{gtf}{GTF file contents for mapping SNPs to genes}

\item{mono}{Rate of mono-allelic expression. Default: 0.7}

\item{pe}{Effective error rate to capture error from sequencing, etc. Default: 0.01}

\item{filter}{Boolean for whether to filter out SNP sites with no coverage. Default: TRUE}

\item{likelihood}{Boolean for whether to use likelihood based estimate of posterior. Default: FALSE}

\item{n.iter}{Number of iterations in MCMC. Default: 1000}

\item{quiet}{Boolean of whether to suppress progress bar. Default: TRUE}

\item{delim}{Delimiter for names of SNPs as Chromosome[delim]Position. Default: ":" ex. chr1:283838897}
}
\value{
Posterior probability of deletion or LOH
}
\description{
Run BADGER allele-only model to assess posterior probability of CNVs given allele information only
}
\examples{
## Single cell data
data(snpsHet_MM16ScSample)
## Bulk exome
data(snpsHet_MM16BulkSample)
\dontrun{
gtfFile <- 'data-raw/Homo_sapiens.GRCh37.75.gtf'
gtf <- read.table(gtfFile, header=F, stringsAsFactors=F, sep='\\t')
}
region <- data.frame('chr'=2, start=0, end=1e9) # deletion region
\dontrun{
results <- calcAlleleCnvProb(r, cov.sc, l, cov.bulk, region, gtf)
}
region <- data.frame('chr'=3, start=0, end=1e9) # neutral region
\dontrun{
results <- calcAlleleCnvProb(r, cov.sc, l, cov.bulk, region, gtf)
}
}

