% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combinedModel.R
\name{calcCombCnvProb}
\alias{calcCombCnvProb}
\title{Run BADGER combined model to assess posterior probability of CNVs given both allele and expression data}
\usage{
calcCombCnvProb(r, cov.sc, l, cov.bulk, region, gtf, gexp, fits, gos, m,
  filter = TRUE, pe = 0.01, mono = 0.7, n.iter = 1000, quiet = TRUE,
  delim = ":")
}
\arguments{
\item{r}{Matrix of alt allele count in single cells}

\item{cov.sc}{Matrix of coverage in single cells}

\item{l}{Vector of alt allele count in bulk}

\item{cov.bulk}{Vector of coverage in bulk}

\item{region}{Region of interest such as expected CNV boundaries}

\item{gexp}{Normalized gene expression matrix}

\item{fits}{Fit for variance around mean}

\item{gos}{Gene position table}

\item{m}{Expected mean deviation due to copy number change}

\item{filter}{Boolean for whether to filter out SNP sites with no coverage. Default: TRUE}

\item{pe}{Effective error rate to capture error from sequencing, etc. Default: 0.01}

\item{mono}{Rate of mono-allelic expression. Default: 0.7}

\item{n.iter}{Number of iterations in MCMC. Default: 1000}

\item{quiet}{Boolean of whether to suppress progress bar. Default: TRUE}

\item{delim}{Delimiter for names of SNPs as Chromosome[delim]Position. Default: ":" ex. chr1:283838897}

\item{likelihood}{Boolean for whether to use likelihood based estimate of posterior. Default: FALSE}
}
\value{
List of posterior probabilities for CNV and direction of CNV (deletion vs. amplification)
}
\description{
Run BADGER combined model to assess posterior probability of CNVs given both allele and expression data
}
\examples{
data('MM16.counts')
mat <- log2(MM16.counts + 1)
data('Normal.counts')
mat.ref <- log2(Normal.counts + 1)
mats <- normalizedExpression(mat, mat.ref)
gexp <- mats[[1]]
fits <- mvFit(gexp)
region <- data.frame('chr'=1, start=0, end=1e9)
data(snpsHet_MM16ScSample)
data(snpsHet_MM16BulkSample)
\dontrun{
gtfFile <- 'data-raw/Homo_sapiens.GRCh37.75.gtf'
gtf <- read.table(gtfFile, header=F, stringsAsFactors=F, sep='\\t')
region <- data.frame('chr'=2, start=0, end=1e9) # deletion region
library(biomaRt) ## for gene coordinates
mart.obj <- useMart(biomart = "ENSEMBL_MART_ENSEMBL", dataset = 'hsapiens_gene_ensembl', host = "jul2015.archive.ensembl.org")
gos <- getBM(values=rownames(mat.tot),attributes=c("ensembl_gene_id","chromosome_name","start_position","end_position"),filters=c("ensembl_gene_id"),mart=mart.obj)
gos$pos <- (gos$start_position + gos$end_position)/2
results <- calcCombCnvProb(r, cov.sc, l, cov.bulk, region, gtf, gexp, fits, gos, m=0.15)
}
}

