% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate.R
\name{calcErrorRate}
\alias{calcErrorRate}
\title{Calculate effective error rate}
\usage{
calcErrorRate(type, snpsCov, snpsRef, snpsAlt, cnvs = NULL, n = 0.1)
}
\arguments{
\item{type}{Type of SNPs provided in \code{snpsCov}, \code{snpsRef}, and \code{snpsAlt} used to estimate effective error. This should be (an unambiguous abbreviation of) one of 'homo_ref', 'homo_alt', or 'het'.}

\item{snpsCov}{A matrix of snp positions as rows and cells as columns with each value corresponding to the total read count at the corresponding snp position in the corresponding cell}

\item{snpsRef}{A matrix of snp positions as rows and cells as columns with each value corresponding to the reference read count at the corresponding snp position in the corresponding cell. Should have the same snp positions and cells as in \code{snpsCov}}

\item{snpsAlt}{A matrix of snp positions as rows and cells as columns with each value corresponding to the alternative read count at the corresponding snp position in the corresponding cell. Should have the same snp positions and cells as in \code{snpsCov}}

\item{cnvs}{A vector of copy number variation regions or other regions with entry as a string in format 'chromosome:start:end'. Optional. Default = NULL}

\item{n}{Minimum fraction of cell in which we must have coverage at a snp position. Useful for eliminating false positives. Default = 0.1}
}
\value{
Effective error rate
}
\description{
Calculate effective error rate from sequencing, alignment, and other potential error sources
}
\examples{
\dontrun{
deletion.region <- data.frame('chr'=2, start=0, end=1e9)
calcErrorRate('het', snpsCov, snpsRef, snpsAlt, cnvs=deletion.region) # ignore regions imapcted by deletion
calcErrorRate('het', snpsCov, snpsRef, snpsAlt) # use all regions
}
}

