% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressionModel.R
\name{calcGexpProb}
\alias{calcGexpProb}
\title{Run BADGER expression-only model to assess posterior probability of CNVs given normalized expression data only}
\usage{
calcGexpProb(region, gexp, fits, m, quiet = TRUE)
}
\arguments{
\item{region}{Region of interest such as expected CNV boundaries}

\item{gexp}{Normalized gene expression matrix}

\item{fits}{Fit for variance around mean}

\item{m}{Expected mean deviation due to copy number change}

\item{quiet}{Boolean for whether to suppress progress display}
}
\value{
List of posterior probabilities for amplification and deletion
}
\description{
Run BADGER expression-only model to assess posterior probability of CNVs given normalized expression data only
}
\examples{
data('MM16.counts')
mat <- log2(MM16.counts + 1)
data('Normal.counts')
mat.ref <- log2(Normal.counts + 1)
mats <- normalizedExpression(mat, mat.ref)
gexp <- mats[[1]]
fits <- mvFit(gexp)
region <- data.frame('chr'=chr, start=0, end=1e9)
set.seed(0)
results <- calcGexpProb(region, gexp, fits, 0.15)
}

