% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpModel.R
\name{clafProfile}
\alias{clafProfile}
\title{Composite Minor Allele Frequency (CLAF) Profile plot}
\usage{
clafProfile(r, n.sc, l, n.bulk, region = NULL, filter = TRUE, delim = ":",
  gtf = NULL, plotGene = FALSE)
}
\arguments{
\item{r}{alt allele count in single cells}

\item{l}{alt allele count in bulk}

\item{cov.sc}{coverage in single cells}

\item{cov}{coverage in bulk}
}
\description{
Composite Minor Allele Frequency (CLAF) Profile plot
}
\examples{
## Single cell data
data(snpsHet_MM16ScSample)
## Bulk exome
data(snpsHet_MM16BulkSample)
# intersect
region <- data.frame('chr'=2, start=0, end=1e9) # deletion region
clafProfile(r, cov.sc, l, cov.bulk, region)
region <- data.frame('chr'=3, start=0, end=1e9) # neutral region
clafProfile(r, cov.sc, l, cov.bulk, region)
}

