% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_BSseq.R
\name{construct_BSseq}
\alias{construct_BSseq}
\title{Create the BSseq object that DSS needs}
\usage{
construct_BSseq(chr, M, Cov)
}
\arguments{
\item{chr}{a string representing which chromosome (e.g. chr18)}

\item{M}{a data.table of methylated read counts with columns 'chromStart', 'sample1', 'sample2', ...}

\item{Cov}{a data.table of coverage with columns 'chromStart', 'sample1', 'sample2', ...}
}
\description{
Wrapper around \code{BSseq()} function that handles the positions,
and M/Cov values
}
\examples{
construct_BSseq()
}
\keyword{BS}
