% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{.generate_nsig_label}
\alias{.generate_nsig_label}
\title{Makes volcano plot}
\usage{
.generate_nsig_label(dmps, alpha)
}
\arguments{
\item{dmps}{a dataframe with a \code{chr}, \code{pos}, \code{x}, \code{y} (\code{y} should be -log10-ed)}

\item{alpha}{a -log10-ed significance level applied to \code{y}}
}
\description{
Creates the skeleton of the Volcano plot (i.e. scatter). You'll need to
customize labels, legend, etc. after the fact. Cannot know a priori what the
axis limits will be.
}
\examples{
plot_volcano()
}
\keyword{plot}
\keyword{viz}
\keyword{volcano}
