% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dss_models.R
\name{fit_dss_models}
\alias{fit_dss_models}
\title{Makes volcano plot}
\usage{
fit_dss_models(bs, df, smoothing, smoothing.span, dss.formula)
}
\arguments{
\item{bs}{a class bisulfite sequence object (see Hansen BSseq)}

\item{dss.formula}{a \code{formula()} representing the model to fit}

\item{design.df}{a data.frame with covariates}

\item{test.var}{a string representing the}
}
\description{
Filters and checks that the design.df and bs samples are identical
Runs DSS models (See Wu 2016)
}
\examples{
fit_dss_models()

}
\keyword{model}
