% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{geneFactors}
\alias{geneFactors}
\title{Returns a factor vector across pos (chr:pos vector)
indicating if the pos is in one of the genes (row number)}
\usage{
geneFactors(posList, gtf, fill = TRUE, gene = TRUE)
}
\arguments{
\item{posList}{Data frame of SNP positions}

\item{gtf}{GTF file contents}

\item{fill}{Boolean of whether to map SNPs not annotated as within genes to unique gene ids}

\item{gene}{Boolean of whether to use genes (or exons only)}
}
\value{
a factor vector across pos (chr:pos vector) indicating if the pos is in one of the genes (row number)
}
\description{
Returns a factor vector across pos (chr:pos vector)
indicating if the pos is in one of the genes (row number)
}
\examples{
posList <- data.frame(
    'chr' = c('1', '1', '1', '2', '2', '3'),
    'pos' = c(11869+7, 11869+100, 14363+101, 29554+0, 52473+10, 62948+5)
    )
\dontrun{
file <- 'data-raw/Homo_sapiens.GRCh37.75.gtf'
gtf <- read.table(gtfFile, header=F, stringsAsFactors=F, sep='\\t')
geneFactors(posList, gtf)
}
}

