% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitaward.R
\name{gitaward}
\alias{gitaward}
\title{gitaward}
\usage{
gitaward(user = "guangchuangyu", lang)
}
\arguments{
\item{user}{user}

\item{lang}{language}
}
\value{
rank
}
\description{
git award rank
}
\examples{
gitaward('guangchuangyu', 'r')
}
\author{
Guangchuang Yu
}

