% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{insideCnvs}
\alias{insideCnvs}
\title{Returns a vector of T/F depending on whether SNPs are in CNV regions}
\usage{
insideCnvs(posList, cnvs)
}
\arguments{
\item{posList}{Dataframe ($chr / $pos)}

\item{cnvs}{Dataframe ($chr / $start / $end)}
}
\value{
a T/F vector of whether the positions are in the CNVs
}
\description{
Returns a vector of T/F depending on whether SNPs are in CNV regions
}
\examples{
snpList <- data.frame(
    'chr' = c('chr1', 'chr1', 'chr1', 'chr2', 'chr2', 'chr3'),
    'pos' = c(7, 100, 101, 0, 10, 5)
    )
cnv <- data.frame(
    'chr' = c('chr1', 'chr2', 'chr3'),
    'startPos' = c(5, 0, 1),
    'endPos' = c(100, 20000, 2)
    )
insideCnvs(snpList, cnv)
}

