% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{insideRange}
\alias{insideRange}
\title{Determine whether a set of SNPs are in a CNV region or other ranges
Super slow...see insideCNVs instead}
\usage{
insideRange(posList, range)
}
\arguments{
\item{posList}{data frame ($chr / $pos)}

\item{cnvs}{: data frame ($chr / $start / $end)}
}
\value{
a T/F vector of whether the positions are in the CNVs
}
\description{
Determine whether a set of SNPs are in a CNV region or other ranges
Super slow...see insideCNVs instead
}
\examples{
snpList <- data.frame(
    'chr' = c('chr1', 'chr1', 'chr1', 'chr2', 'chr2', 'chr3'),
    'pos' = c(7, 100, 101, 0, 10, 5)
    )
cnv <- data.frame(
    'chr' = c('chr1', 'chr2', 'chr3'),
    'startPos' = c(5, 0, 1),
    'endPos' = c(100, 20000, 2)
    )
insideRanges(snpList, cnv)
}

