% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressionModel.R
\name{mvFit}
\alias{mvFit}
\title{Estimate expression magnitude variance as a function of number of genes}
\usage{
mvFit(gexp, num.genes = seq(5, 100, by = 5), rep = 1000, plot = FALSE)
}
\arguments{
\item{gexp}{Normalized gene expression matrix, preferably restricted to region of known neutral copy number}

\item{num.genes}{List of number of genes for random sampling. Default: seq(5, 100, by=5)}

\item{rep}{Number of replicates. Default: 1000}

\item{plot}{Boolean of whether to plot. Default: FALSE}
}
\value{
Fit for variance around mean normalized gene expression as a function of the number of genes for each cell
}
\description{
Estimate expression magnitude variance as a function of number of genes
}
\examples{
data('MM16.counts')
mat <- log2(MM16.counts + 1)
data('Normal.counts')
mat.ref <- log2(Normal.counts + 1)
mats <- normalizedExpression(mat, mat.ref)
gexp <- mats[[1]]
fits <- mvFit(gexp)
}

