% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressionModel.R
\name{normalizedExpression}
\alias{normalizedExpression}
\title{Normalize gene expression to a normal reference for expression-based karyotyping}
\usage{
normalizedExpression(mat, mat.ref, minMeanBoth = 4.5, minMeanTest = 6,
  minMeanRef = 8, verbose = TRUE)
}
\arguments{
\item{mat}{Test gene expression matrix to be normalized}

\item{mat.ref}{Normal reference gene expression matrix}

\item{minMeanBoth}{Filter out genes with mean gene expression below this threshold in both samples. Default: 4.5}

\item{minMeanTest}{Filter out genes with mean gene expression below this threshold in the test gene expression matrix. Default: 6}

\item{minMeanRef}{Filter out genes with mean gene expression below this threshold in the reference gene expression matrix. Default: 8}

\item{verbose}{Boolean for whether to print out verbose warnings. Default: TRUE}
}
\value{
List of two normalized gene expression matrices corresponding to mat and mat.ref
}
\description{
Normalize gene expression to a normal reference for expression-based karyotyping
}
\examples{
data('MM16.counts')
mat <- log2(MM16.counts + 1)
data('Normal.counts')
mat.ref <- log2(Normal.counts + 1)
mats <- normalizedExpression(mat, mat.ref)
}

