% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressionModel.R
\name{plotExpHeatmap}
\alias{plotExpHeatmap}
\title{Visualize expression heatmap by chromosome for expression-based karyotyping}
\usage{
plotExpHeatmap(mat.tot, gos, zlim = c(-2, 2), window.size = 101,
  autosomesOnly = TRUE, orderCells = FALSE)
}
\arguments{
\item{mat.tot}{Gene expression matrix}

\item{gos}{Dataframe of positional information for each gene in the gene expression matrix}

\item{zlim}{Lower and upper bounds for heatmap. Default: c(-2,2)}

\item{window.size}{Window size for sliding window average. Default: 101}

\item{autosomesOnly}{Boolean for whether to plot only autosomes. Default: TRUE, if false will also plot X chromosome}

\item{orderCells}{Booean for whether to automaticaly cluster and order cells. Default: FALSE}
}
\value{
List of sliding window averaged expression by chromosome along with heatmap representation
}
\description{
Visualize expression heatmap by chromosome for expression-based karyotyping
}
\examples{
data('MM16.counts')
mat <- log2(MM16.counts + 1)
data('Normal.counts')
mat.ref <- log2(Normal.counts + 1)
mats <- normalizedExpression(mat, mat.ref)
mat.tot <- cbind(mats[[1]], mats[[2]])
library(biomaRt) ## for gene coordinates
mart.obj <- useMart(biomart = "ENSEMBL_MART_ENSEMBL", dataset = 'hsapiens_gene_ensembl', host = "jul2015.archive.ensembl.org")
gos <- getBM(values=rownames(mat.tot),attributes=c("ensembl_gene_id","chromosome_name","start_position","end_position"),filters=c("ensembl_gene_id"),mart=mart.obj)
gos$pos <- (gos$start_position + gos$end_position)/2
tl <- plotExpHeatmap(mat.tot, gos, zlim=c(-0.5,0.5), window.size = 201)
}

