% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{pointsWithin}
\alias{pointsWithin}
\title{Determine if a point is within a range}
\usage{
pointsWithin(pos, ranges)
}
\arguments{
\item{pos}{Dataframe with chr and pos columns}

\item{ranges}{Dataframe with chr, start, and end columns}
}
\description{
Determine if a point is within a range
}
\examples{
snp1 <- data.frame(
    'chr' = 'chr1',
    'pos' = 7
    )
cnv <- data.frame(
    'chr' = c('chr1', 'chr2', 'chr3'),
    'startPos' = c(5, 0, 1),
    'endPos' = c(100, 20000, 2)
    )
pointsWithin(snp1, cnv)

# no hits
snp2 <- data.frame(
    'chr' = 'chrX',
    'pos' = 7
    )
pointsWithin(snp2, cnv)
}

