% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{pos2GRanges}
\alias{pos2GRanges}
\title{From data frame of positions to GRanges object}
\usage{
pos2GRanges(df)
}
\arguments{
\item{df}{Dataframe with chr and pos columns}
}
\description{
From data frame of positions to GRanges object
}
\examples{
snps <- data.frame(
'chr' = c('chr2', 'chr1'),
   'pos' = c(20005, 7)
   )
pos2GRanges(snps)
}

