% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSnps.R
\name{process.bulk}
\alias{process.bulk}
\title{Process MuTect output from WES}
\usage{
process.bulk(snps.file, out.file)
}
\arguments{
\item{snps.file}{MuTect .mut file}

\item{out.file}{RData output file location and name}
}
\value{
outputs positions and allele information for somatic, germline het, germline homo ref and germline homo alt snps
 saves output to out.file
}
\description{
Process MuTect output from WES
}

