% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsWithin.R
\name{range2GRanges}
\alias{range2GRanges}
\title{From data frame of ranges to GRanges object}
\usage{
range2GRanges(df)
}
\description{
From data frame of ranges to GRanges object
}
\examples{
cnv <- data.frame(
   'chr' = c('chr1', 'chr2'),
   'startPos' = c(5, 0),
   'endPos' = c(100, 20000)
   )
range2GRanges(df)
}

