% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_altmetric}
\alias{badge_altmetric}
\title{badge_altmetric}
\usage{
badge_altmetric(id, color)
}
\arguments{
\item{id}{altmetric id}

\item{color}{color of badge}
}
\value{
badge in markdown syntax
}
\description{
altmetric badge
}
\author{
Guangchuang
}

