% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_bioc_download}
\alias{badge_bioc_download}
\title{badge_bioc_download}
\usage{
badge_bioc_download(pkg, by, color, type = "distinct")
}
\arguments{
\item{pkg}{package}

\item{by}{one of total or month}

\item{color}{badge color}

\item{type}{one of distinct and total}
}
\value{
badge in markdown syntax
}
\description{
badge of bioconductor download number
}
\author{
Guangchuang Yu
}

