% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_bioc_release}
\alias{badge_bioc_release}
\title{badge_bioc_release}
\usage{
badge_bioc_release(pkg, color)
}
\arguments{
\item{pkg}{package}

\item{color}{badge color}
}
\value{
badge in markdown syntax
}
\description{
badge of bioconductor release version
}
\author{
Guangchuang Yu
}
