% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_code_size}
\alias{badge_code_size}
\title{badge_code_size}
\usage{
badge_code_size(ref)
}
\arguments{
\item{ref}{Reference for a GitHub repository}
}
\value{
badge in markdown syntax
}
\description{
badge of GitHub code size
}
\author{
Gregor de Cillia
}
