% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_codecov}
\alias{badge_codecov}
\title{badge_codecov}
\usage{
badge_codecov(ref = NULL, token = NULL, branch = "master")
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the DESCRIPTION file.}

\item{token}{Codecov graphing token, needed for private repositories.
It can be obtained at https://codecov.io/gh/USER/REPO/branch/BRANCH/graph/}

\item{branch}{The GitHub branch. Defaults to \code{"master"}.}
}
\value{
badge in markdown syntax
}
\description{
badge of codecov code coverage
}
\author{
Gregor de Cillia
}
