% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_coveralls}
\alias{badge_coveralls}
\title{badge_coveralls}
\usage{
badge_coveralls(ref)
}
\arguments{
\item{ref}{Reference for a GitHub repository}
}
\value{
badge in markdown syntax
}
\description{
badge of CRAN release version
}
\author{
Gregor de Cillia
}
