% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_cran_checks}
\alias{badge_cran_checks}
\title{badge_cran_checks}
\usage{
badge_cran_checks(pkg = NULL)
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}
}
\value{
badge in Markdown syntax
}
\description{
CRAN checks badge
}
\author{
Scott Chamberlain (badges API), Maëlle Salmon (function)
}
