% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_cran_download}
\alias{badge_cran_download}
\title{badge_cran_download}
\usage{
badge_cran_download(pkg)
}
\arguments{
\item{pkg}{package}
}
\value{
badge in markdown syntax
}
\description{
badge of CRAN downloads
}
\author{
Gregor de Cillia
}
