% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_cran_download}
\alias{badge_cran_download}
\title{badge_cran_download}
\usage{
badge_cran_download(
  pkg = NULL,
  type = c("last-month", "last-week", "grand-total"),
  color = "green"
)
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}

\item{type}{type of stats. \code{last-month}, \code{last-week} or \code{"grand-total"}}

\item{color}{color of badge}
}
\value{
badge in markdown syntax
}
\description{
badge of CRAN downloads
}
\author{
Gregor de Cillia
}
