% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_cran_release}
\alias{badge_cran_release}
\title{badge_cran_release}
\usage{
badge_cran_release(pkg, color)
}
\arguments{
\item{pkg}{package}

\item{color}{color of badge}
}
\value{
badge in markdown syntax
}
\description{
badge of CRAN version
}
\author{
Gregor de Cillia
}
