% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_dependencies}
\alias{badge_dependencies}
\title{badge_depedencies}
\usage{
badge_dependencies(pkg = NULL)
}
\arguments{
\item{pkg}{package. If \code{NULL} (the default) the package
is determined via the DESCRIPTION file.}
}
\value{
badge in markdown syntax
}
\description{
dependency badge
}
\author{
Dirk Eddelbuettel
}
