% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_dependencies}
\alias{badge_dependencies}
\title{badge_depedencies}
\usage{
badge_dependencies(pkg)
}
\arguments{
\item{pkg}{package}
}
\value{
badge in markdown syntax
}
\description{
dependency badge
}
\author{
Dirk Eddelbuettel
}
