% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_devel}
\alias{badge_devel}
\title{badge_devel}
\usage{
badge_devel(pkg, color)
}
\arguments{
\item{pkg}{package}

\item{color}{badge color}
}
\value{
badge in markdown syntax
}
\description{
badge of devel version
}

