% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_doi}
\alias{badge_doi}
\title{badge_doi}
\usage{
badge_doi(doi, color)
}
\arguments{
\item{doi}{doi}

\item{color}{color}
}
\value{
badge
}
\description{
doi badge
}
\examples{
badge_doi("10.1111/2041-210X.12628", "green")
}
\author{
Guangchuang
}

