% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_github_release}
\alias{badge_github_release}
\title{GitHub Release badge}
\usage{
badge_github_release(ref = NULL)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL} (the default),
the reference is determined by the URL field in the DESCRIPTION file.}
}
\value{
badge in markdown syntax
}
\description{
GitHub Release badge
}
\author{
Matt Schuelke
}
