% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_github_version}
\alias{badge_github_version}
\title{badge_github_version}
\usage{
badge_github_version(pkg, color)
}
\arguments{
\item{pkg}{package}

\item{color}{badge color}
}
\value{
badge in markdown syntax
}
\description{
badge of devel version
}
\author{
Guangchuang Yu
}

