% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_last_commit}
\alias{badge_last_commit}
\title{badge_last_commit}
\usage{
badge_last_commit(ref = NULL)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the description file.}
}
\value{
badge in markdown syntax
}
\description{
last commit badge
}
\author{
Gregor de Cillia
}
