% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_lifecycle}
\alias{badge_lifecycle}
\title{badge_lifcycle}
\usage{
badge_lifecycle(status = "experimental", color)
}
\arguments{
\item{status}{lifecycle status. See \link{https://www.tidyverse.org/lifecycle/}}

\item{color}{color of badge}
}
\value{
badge in markdown syntax
}
\description{
lifecycle badge
}
\author{
Gregor de Cillia
}
