% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_runiverse}
\alias{badge_runiverse}
\title{r-universe badge}
\usage{
badge_runiverse(pkg = NULL, user = NULL)
}
\arguments{
\item{pkg}{Package name. If \code{NULL} (the default),
the name is determined by the URL field in the DESCRIPTION file.}

\item{user}{User name. If \code{NULL} (the default),
the name is determined by the URL field in the DESCRIPTION file.}
}
\value{
badge in markdown syntax
}
\description{
r-universe badge
}
\author{
Alexander Rossell Hayes
}
