% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_travis}
\alias{badge_travis}
\title{badge_travis}
\usage{
badge_travis(ref)
}
\arguments{
\item{ref}{Reference for a GitHub repository}
}
\value{
badge in markdown syntax
}
\description{
reavis-ci badge
}
\author{
Gregor de Cillia
}
