% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_travis}
\alias{badge_travis}
\title{badge_travis}
\usage{
badge_travis(ref = NULL)
}
\arguments{
\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the DESCRIPTION file.}
}
\value{
badge in markdown syntax
}
\description{
reavis-ci badge
}
\author{
Gregor de Cillia
}
