% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger-package.R
\docType{package}
\name{badger-package}
\alias{badger}
\alias{badger-package}
\title{badger: Badge for R Package}
\description{
Query information and generate badge for using in README and GitHub Pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GuangchuangYu/badger}
  \item Report bugs at \url{https://github.com/GuangchuangYu/badger/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Gregor de Cillia \email{de.cillia.gregor@gmail.com} [contributor]
  \item Dirk Eddelbuettel \email{edd@debian.org} [contributor]
  \item Maëlle Salmon [contributor]
  \item Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood) [contributor]
  \item Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) (rossellhayes) [contributor]
  \item Waldir Leoncio \email{w.l.netto@medisin.uio.no} [contributor]
}

}
\keyword{internal}
