% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{ver_devel}
\alias{ver_devel}
\title{ver_devel}
\usage{
ver_devel(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
devel version
}
\description{
get devel version number of specific package
}
\author{
Guangchuang
}
